third_party/WebKit/LayoutTests/fast/html/navigation-transition-expected.txt

Tests for the HTML markup serializer used for transition elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS sourceElementRect.top is destElement.offsetTop
PASS sourceElementRect.left is destElement.offsetLeft
PASS sourceElementRect.height is destElement.offsetHeight
PASS sourceElementRect.width is destElement.offsetWidth
PASS preHidingTransitionElementOpacity == 1 is true
PASS postHidingTransitionElementOpacity == 0 is true
PASS preHidingBottomElementRect.top is postHidingBottomElementRect.top
PASS preHidingBottomElementRect.left is postHidingBottomElementRect.left
PASS successfullyParsed is true

TEST COMPLETE
Header

Footer
